# src/main/swagger/temowl_retrieval.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: "SAM, the semantic aggregation machine"
  name: SAM
  description: >-
              API to retrieve and export Terminologies and their contents.
              Some API Calls expect and/or return a TerminologyID. This ID is configured by the System and depends on the
              system configuration. A list of all available Terminologies can be obtained via the API Call
              `\terminologies`.
              Even though the ID is defined by the system reuse of IDs is intended. For example the TerminologyID of the
              Terminology 'Gene Ontology' should be 'GO'.
              If a Terminology is missing or uses an uncommon TerminologyID please contact the system administrator.
              Issued API Calls with a not present TerminologyID will lead to an error.

              Some API Calls expect and/or return a ConceptID. The ID is either encoded as CURIE
              (https://www.w3.org/TR/2010/NOTE-curie-20101216/) or a canonical IRI (PURL?!).
              A CURIE encoding is used if the originating Terminology (the one that initial defined this Concept) is also present
              in this System, otherwise an IRI is used. However requests issued with an IRI are redirected to a CURIE if
              possible.
              However as multiple IRI could be defined for a single Concept this redirect may not work. Please contact the
...

...

...
              the System.

...
              cases.
...

...

...

...

...


...
basePath: /api/v1
produces:
...
parameters:
    sizeParam:
        in: query
        name: size
        description: Size of pages
        type: integer
        format: int64
        default: 10
    pageParam:
        in: query
        name: page
        description: The requested page
        type: integer
        format: int64
        default: 0
    langParam:
        in: query
        name: lang
        description: This allows to filter the response to only contain descriptions in a
                 specific language. If no description in the given language is present non is returned. The API
                 expect a language to be encoded as defined in https://tools.ietf.org/html/bcp47. Further an
                 extension 'all' is allowed. If that is given properties in all language are returned. The default
                 language is 'en'.
        default: 'en'
        type: string
    versionTagParam:
        in: query
        name: versionTag
        description: Terminology version defined by tag-name
        type: string
...
    terminologyIDParam:
        in: path
        name: terminologyId
        description: Identifier of the terminology.
        required: true
        type: string
...
responses:
    NotFound:
      description: Entity not found.
      schema:
        $ref: '#/definitions/Error'
    BadClient:
      description: Bad Client Request
      schema:
        $ref: '#/definitions/Error'
...
schemes:
...
paths:
...
  '/terminologies/{terminologyId}/versionTags':
      get:
        tags:
          - retrieval
        summary: Retrieve all versions of a Terminology.
        description: Returns all version tags of the specified terminology, paginated using page and size.
        operationId: Tags-view
        parameters:
          - $ref: '#/parameters/terminologyIDParam'
          - $ref: '#/parameters/pageParam'
          - $ref: '#/parameters/sizeParam'
        responses:
            '200':
                description: Successful response
                schema:
                  $ref: '#/definitions/PagedVersion'
            '400':
                $ref: '#/responses/BadClient'
            '404':
                $ref: '#/responses/NotFound'
  '/terminologies/':
    get:
      tags:
        - retrieval
      summary: Get all terminologies
      description: >-
        Returns a terminology's meta information, including a link to the original source and a link to the local refined version, if present. Also returns links to changes and versionTags.
      operationId: Terminology-Get
      parameters:
        - $ref: '#/parameters/langParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/sizeParam'
      responses:
        '200':
          description: Successful response
          schema:
           $ref: '#/definitions/PagedTerminology'
        '400':
            $ref: '#/responses/BadClient'
        '404':
            $ref: '#/responses/NotFound'
  '/terminologies/{terminologyId}/':
    get:
      tags:
        - retrieval
      summary: Get a terminology's meta information
      description: >-
        Returns a terminology's meta information, including a link to the original source and a link to the local refined version, if present. Also returns links to changes and versionTags.
      operationId: Terminology-Get-Meta
      parameters:
          - $ref: '#/parameters/terminologyIDParam'
          - $ref: '#/parameters/versionTagParam'
          - $ref: '#/parameters/langParam'
      responses:
        '200':
          description: Successful response
          schema:
           $ref: '#/definitions/Terminology'
        '400':
            $ref: '#/responses/BadClient'
        '404':
            $ref: '#/responses/NotFound'
...
definitions:
  Terminology:
    type: object
    description: "A Terminology"
    required:
      - terminologyID
      - baseURL
      - shortName
    properties:
      terminologyID:
        description: "A unique ID of the Terminology. Used as a prefix for CURIES, and a Parameter in different API
        calls."
        type: string
        example: 'GO'
      baseURL:
        description: "Base URL of the Terminology can be used to subsitute a CURIE Prefix."
        type: string
        example: 'http://purl.obolibrary.org/obo/GO_'
      longName:
        description: "The Name of the Terminology."
        type: string
        example: 'Gene Ontology'
      descriptions:
       # multiple languages?!?!?
       description: "A Description of the Terminology"
       $ref: '#/definitions/Description'
      citation:
       description: "Article to cite if this Terminology is used."
       type: string
       example: 'https://www.ncbi.nlm.nih.gov/pubmed/10802651'
      versiontags:
        description: "A list of versionTags that are present for this Terminology"
        type: string
        example: 'api.example.org/terminologies/GO/versionTags'
...
  Description:
    type: object
    description: "A `description` in a specific language."
    required:
      - description
      - lang
    properties:
      description:
        description: "A description."
        type: string
        example: 'This is a fairly long description.'
      lang:
        description: "The language of this description, as defined by https://tools.ietf.org/html/bcp47."
        type: string
        example: 'en'
...
  VersionTag:
    type: object
    required:
      - shortName
      - timestamp
    properties:
      shortName:
        type: string
        description: "A short unique meaningfull name of this version. Like a release name."
        example: "etch"
      timestamp:
        type: string
        format: date-time
        description: "The time at which the version tag was created."
        example: "2018-04-15-8:30"
      description:
        description: "A description of a version."
        $ref: '#/definitions/Description'
####################### Other definitions
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
        example: 400
      message:
        type: string
        description: error type
      cause:
        type: string
        description: detailed information on what happened
####################### Pagination extensions
  Page:
    type: object
    required:
      - number
      - numberOfElements
      - totalElements
      - totalPages
      - size
    properties:
        number:
          type: integer
          description: ' Number of this page'
          example: "0"
        numberOfElements:
          type: integer
          description: ' Number of elements in this page'
          example: "4"
        totalElements:
          type: integer
          description: 'total Number of elements'
          example: "4"
        totalPages:
          type: integer
          description: 'total Number of page'
          example: "1"
        size:
          type: integer
          description: 'size of a page'
          example: "10"
...
  PagedTerminology:
    allOf:
        - $ref: '#/definitions/Page'
        - type: object
          required:
           - content
          properties:
            content:
                type: array
                items:
                 $ref: '#/definitions/Terminology'
...
  PagedVersion:
    allOf:
        - $ref: '#/definitions/Page'
        - type: object
          required:
           - content
          properties:
            content:
                type: array
                items:
                  $ref: '#/definitions/VersionTag'
####################### BEL extensions
...